+++ src/tools/timings.f90
@@ -16,8 +16,7 @@
 
       INTEGER, INTENT( OUT ) :: status
       CHARACTER ( LEN = * ), INTENT( IN ) :: name
-      REAL ( KIND = wp ), INTENT( out ) :: time
-
+      REAL, INTENT( out ) :: time
 !  ---------------------------------------------------------------------------
 !  return the total CPU time spent in the cutest evaluation tool called 'name'
 !  while the CPU monitor was turned on (see cutest_set_monitor). Timings are
@@ -51,7 +50,7 @@
       INTEGER, INTENT( IN ) :: thread
       INTEGER, INTENT( OUT ) :: status
       CHARACTER ( LEN = * ), INTENT( IN ) :: name
-      REAL ( KIND = wp ), INTENT( out ) :: time
+      REAL, INTENT( out ) :: time
 
 !  ---------------------------------------------------------------------------
 !  return the total CPU time spent in the cutest evaluation tool called 'name'
@@ -98,7 +97,7 @@
       TYPE ( CUTEST_work_type ) :: work
       INTEGER, INTENT( OUT ) :: status
       CHARACTER ( LEN = * ), INTENT( IN ) :: name
-      REAL ( KIND = wp ), INTENT( out ) :: time
+      REAL, INTENT( out ) :: time
 
 !  ---------------------------------------------------------------------------
 !  return the total CPU time spent in the cutest evaluation tool called 'name'
@@ -111,10 +110,10 @@
       SELECT CASE ( name )
       CASE ( 'start' )
         work%record_times = .TRUE.
-        time = 0.0_wp
+        time = 0.0
       CASE ( 'stop' )
         work%record_times = .FALSE.
-        time = 0.0_wp
+        time = 0.0
       CASE ( 'cutest_ccfg' )
         time = work%time_ccfg
       CASE ( 'cutest_ccfsg' )
@@ -211,7 +210,7 @@
         time = work%time_ushprod
       CASE DEFAULT
         status = 26
-        time = 0.0_wp
+        time = 0.0
         IF ( data%out > 0 ) WRITE( data%out,                                   &
           "( ' ** SUBROUTINE TIMINGS: unknown evaluation function ', A )" )    &
             TRIM ( name )
